Endian conversion for stored ByteChunks on OSX PPC
Make it so a project stored on a PPC Mac can be opened
on an Intel Mac (and vice versa).
Data is always saved in little endian format.
No change to the intel code, but in the PPC compile
functions are added to convert int's and double's
(the two data types IPlug uses internally) to
little endian before writing and from little
endian when reading.